Puliendo el programa

+++ Entities/SaleDetail.cs
@@ -6,7 +6,7 @@ namespace Store.Entities
         public Almacen Store { get; set; }
         public Producto Product { get; set; }
         public int Cantidad { get; set; }
-        public int Descuento { get; set; }
+        public decimal Descuento { get; set; }
         public decimal CostoUnitario { get; set; }
         public decimal PVM { get; set; }
         public decimal PVD { get; set; }
